Extendable typings (#3477)
## Proposed changes

[//]: # (Describe the big picture of your changes here to communicate to the maintainers why we should accept this pull request. If it fixes a bug or resolves a feature request, be sure to link to that issue.)

Proposal for adding types to frameworks/services etc...
Please have a look at the proposed changes here and let me know what you think. It does 2 things:
- The current typings make no difference in the `suite` and `test` description, regardless of framework. Actually, they are currently describing the mocha output. However, when running jasmine, there is different output available. With this PR all you would have to do is reference the framework package in your `tsconfig.json` and it would give you the correct typings for `suite` and `test` AND give you global typings for either mocha or jasmine.
- As an example I also added typings in this way for the `selenium-standalone-service`. To get the correct typings and allowed extra properties to the `config` object, all you need to do is install selenium-standalone-service and add a reference to it in the `tsconfig.json`.

Look at the `typescript.md` file for a description/example. This is just a proposal, but one that would work pretty nicely I think. Ideally, all wdio packages would come with typings files like this.

Please let me know what you think about this.

## Types of changes

[//]: # (What types of changes does your code introduce to WebdriverIO?)
[//]: # (_Put an `x` in the boxes that apply_)

- [ ] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [x] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist

[//]: # (_Put an `x` in the boxes that apply. You can also fill these out after creating the PR. If you're unsure about any of them, don't hesitate to ask. We're here to help! This is simply a reminder of what we are going to look for before merging your code._)

- [x] I have read the [CONTRIBUTING](https://github.com/webdriverio/webdriverio/blob/master/CONTRIBUTING.md) doc
- [ ] I have added tests that prove my fix is effective or that my feature works
- [x] I have added necessary documentation (if appropriate)

## Further comments

[//]: # (If this is a relatively large or complex change, kick off the discussion by explaining why you chose the solution you did and what alternatives you considered, etc...)

### Reviewers: @webdriverio/technical-committee
