+++ .gitignore
@@ -64,6 +64,9 @@ typings/
 /*.js
 *.d.ts
 !scripts/templates/*.d.ts
+!/packages/wdio-mocha-framework/mocha-framework.d.ts
+!/packages/wdio-jasmine-framework/jasmine-framework.d.ts
+!/packages/wdio-selenium-standalone-service/selenium-standalone-service.d.ts
 !babel.config.js
 !.eslintrc.js
 package-lock.json
