+++ docs/TypeScript.md
@@ -41,3 +41,50 @@ and your `tsconfig.json` needs to look like:
     ]
 }
 ```
+
+You can even use a typed configuration if you desire.
+All you have to do is create a plain js config file that registers typescript and requires the typed config:
+
+```javascript
+require("ts-node/register")
+module.exports = require("wdio.conf.ts")
+```
+
+And in your typed configuration file:
+
+```typescript
+const config: WebdriverIO.Config = {
+    // Put your webdriverio configuration here
+}
+
+export { config }
+```
+
+Depending on the framework you use, you will need to add the typings for that framework to your `tsconfig.json` types property.
+For instance, if we decide to use the mocha framework, we need to add it like this to have all typings globally available:
+
+```json
+{
+    "compilerOptions": {
+        "baseUrl": ".",
+        "paths": {
+            "*": [ "./*" ],
+            "src/*": ["./src/*"]
+        },
+        "types": ["node", "webdriverio", "@wdio/mocha-framework"]
+    },
+    "include": [
+        "./src/**/*.ts"
+    ]
+}
+```
+
+Instead of having all type definitions globally available, you can also import only the typings that you need like this:
+
+```typescript
+/*
+* These import the type definition for the `test` and `suite` variables that are available in
+* the beforeTest, afterTest, beforeSuite and afterSuite hooks.
+*/
+import { Suite, Test } from "@wdio/mocha-framework"  
+```
