+++ packages/wdio-jasmine-framework/jasmine-framework.d.ts
@@ -1,0 +1,14 @@
+/// <reference types="jasmine"/>
+
+declare namespace WebDriverIO {
+    interface Suite {
+        title: string;
+        fullName: string;
+        file: string;
+    }
+
+    interface Test extends Suite {
+        parent: string;
+        passed: boolean;
+    }
+}
