+++ packages/wdio-jasmine-framework/package.json
@@ -33,6 +33,7 @@
     "url": "https://github.com/webdriverio/webdriverio/issues"
   },
   "dependencies": {
+    "@types/jasmine": "^3.3.8",
     "@wdio/config": "^5.4.6",
     "@wdio/logger": "^5.4.6",
     "jasmine": "^3.2.0"
@@ -42,5 +43,6 @@
   },
   "publishConfig": {
     "access": "public"
-  }
+  },
+  "types": "jasmine-framework.d.ts"
 }
