+++ packages/wdio-mocha-framework/mocha-framework.d.ts
@@ -1,0 +1,16 @@
+/// <reference types="mocha"/>
+
+declare namespace WebDriverIO {
+    interface Suite {
+        title: string;
+        parent: string;
+        fullTitle: string;
+        pending: boolean;
+    }
+
+    interface Test extends Suite {
+        currentTest: string;
+        passed: boolean;
+        duration: any;
+    }
+}
