+++ packages/wdio-mocha-framework/package.json
@@ -30,6 +30,7 @@
     "url": "https://github.com/webdriverio/webdriverio/issues"
   },
   "dependencies": {
+    "@types/mocha": "^5.2.5",
     "@wdio/config": "^5.4.6",
     "@wdio/logger": "^5.4.6",
     "mocha": "^6.0.0-1"
@@ -39,5 +40,6 @@
   },
   "publishConfig": {
     "access": "public"
-  }
+  },
+  "types": "mocha-framework.d.ts"
 }
