+++ packages/wdio-selenium-standalone-service/package.json
@@ -35,6 +35,7 @@
     "url": "https://github.com/webdriverio/webdriverio/issues"
   },
   "dependencies": {
+    "@types/selenium-standalone": "^6.15.0",
     "fs-extra": "^6.0.1",
     "selenium-standalone": "^6.15.1"
   },
@@ -43,5 +44,6 @@
   },
   "publishConfig": {
     "access": "public"
-  }
+  },
+  "types": "selenium-standalone-service.d.ts"
 }
