+++ packages/wdio-selenium-standalone-service/selenium-standalone-service.d.ts
@@ -1,0 +1,10 @@
+import { InstallOpts, StartOpts } from "selenium-standalone";
+
+declare namespace WebdriverIO {
+    interface Config {
+        seleniumLogs?: string;
+        seleniumInstallArgs?: InstallOpts;
+        seleniumArgs?: StartOpts;
+        skipSeleniumInstall?: boolean;
+    }
+}
